get rid of calls subset()
There is a very special case in BWI3_HR_Totholz_Funktionen_v1.R where I need to
set a data.frame of all NA created by subsetting to empty, which is what
subset() returns! Mind that for future errors after eliminating calls to
subset(). Was quite a pain in the ass to figure that out.